packages/pytest-httpchain-models/pyproject.toml
[project]
name = "pytest-httpchain-models"
version = "0.2.1"
description = "Pydantic models for pytest-httpchain"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Alexander Eresov", email = "aeresov@gmail.com" }]
dependencies = [
    "jmespath>=1.0.0",
    "jsonschema>=4.0.0",
    "graphql-core>=3.0.0",
    "pytest-httpchain-templates",
    "pytest-httpchain-userfunc",
]

[build-system]
requires = ["uv_build>=0.7.21,<0.8.0"]
build-backend = "uv_build"
